+++ lab5/fgjyh/app_url.txt
@@ -1,0 +1,1 @@
+https://fgjyh-cal-gcd.herokuapp.com/
